refractor(locker):split-use-locker-actions
